rsp: fix single stepping (again)
In 7de4fae the resume code was changed so that the first instruction after
hitting a BP was re-executed (while the BP was temporarily disabled)
before continuing normally by either stepping or continuing execution.
However, this is not what GDB expects since it takes care to remove
(i.e. to not enable) such break points when stepping anyway.
Instead the old code broke the sequence of single-stepping after
hitting a break point, i.e.:
  br somewhere
  c
  # Hit BP
  stepi
  stepi

While the first step after the BP works fine, the second one (and any
subsequent one!) re-executes the instruction after the BP - forever.

This patch undoes most of 7de4fae that is related to BPs but keeps the
one thing that presumably fixes the problem it was meant to repair:
Not immediately re-executing an instruction hidden by a BP when the
respective BP is removed by a Z packet - again gdb takes care of this.